Fix for swapped text and click events
When cursor is in a GuiTextField and a "button" is clicked (potentially closing the dialogue box down), the click event is triggered (and the dlg box is closed) before the PlayerGuiInputEvent and before saving the text control contents.

This fix stores the "inText" status of a GuiModalWindow and, if set, click events rare ignored until a PlayerGuiInputEvent clears it.